feat: simplify page resolver logic, shared useCms loading and error state
